+++ packages/scripts/src/index.ts
@@ -5,4 +5,6 @@ export * from './node-environments-manager';
 export * from './forked-process';
 export * from './types';
 export * from './utils';
-export * from './run-node-environment';
+export * from './node-environment';
+export * from './ws-environment';
+export * from './ipc-environment';
